# Kafka & Kinesis stream ingest parsing in parallel