--- a/indexing-service/src/main/java/org/apache/druid/indexing/seekablestream/common/OrderedPartitionableRecord.java
+++ b/indexing-service/src/main/java/org/apache/druid/indexing/seekablestream/common/OrderedPartitionableRecord.java
@@ -83,7 +83,7 @@ public class OrderedPartitionableRecord<PartitionIdType, SequenceOffsetType, Rec
    * @return the list of ByteEntity object for this record
    */
   @NotNull
-  public List<? extends ByteEntity> getData()
+  public List<RecordType> getData()
   {
     return data;
   }