--- a/indexing-service/src/test/java/org/apache/druid/indexing/seekablestream/SeekableStreamSupervisorSpecTest.java
+++ b/indexing-service/src/test/java/org/apache/druid/indexing/seekablestream/SeekableStreamSupervisorSpecTest.java
@@ -500,6 +500,7 @@ public class SeekableStreamSupervisorSpecTest extends EasyMockSupport
             null,
             null,
             null,
+            null,
             null
         )
         {