@FrankChen021 2022-03-04
Do you have a performance report that shows how this change improves the throughput by setting different values of thread count under same incoming message rate?