@exherb 2022-06-13
about 36% time spend on parsing.